requirements_minimal.txt
torch
torchvision
opencv-python
librosa
numpy
pandas
matplotlib
seaborn
scikit-learn
python-dotenv